Updated Trips VC
I removed the view controller for Trips and replaced it with a
TableView Controller. The ‘New Trip’ button should be a + in the
navigation bar (i’ve added this).

I’ve also tried to keep with a lighter theme to match Pierces styles.
We should probably chat and see what colour scheme we want to use on
iOS.